# Conflicts:
#	build_webpack/webpack.terminal.config.js
#	src/common/blockchain/interface-blockchain/agents/Nano/Nano-Wallet-Protocol.js
#	src/common/blockchain/interface-blockchain/blockchain/Interface-Blockchain.js
#	src/common/blockchain/interface-blockchain/blocks/Interface-Blockchain-Block-Data-Transactions.js
#	src/common/blockchain/interface-blockchain/blocks/Interface-Blockchain-Block.js
#	src/common/blockchain/interface-blockchain/blocks/Interface-Blockchain-Blocks.js
#	src/common/blockchain/interface-blockchain/free-memory/Interface-Blockchain-Free-Memory.js
#	src/common/blockchain/interface-blockchain/mining/Interface-Blockchain-Mining.js
#	src/common/blockchain/interface-blockchain/mining/transactions-selector/Mining-Transactions-Selector.js
#	src/common/blockchain/interface-blockchain/protocol/Interface-Blockchain-Protocol-Fork-Solver.js
#	src/common/blockchain/interface-blockchain/transactions/pending/Transactions-Pending-Queue.js
#	src/common/blockchain/interface-blockchain/transactions/protocol/Transactions-Download-Manager.js
#	src/common/blockchain/interface-blockchain/transactions/protocol/Transactions-List-For-Propagation.js
#	src/common/blockchain/interface-blockchain/transactions/protocol/Transactions-Protocol.js
#	src/common/blockchain/interface-blockchain/transactions/transaction/Interface-Blockchain-Transaction.js
#	src/common/blockchain/ppow-blockchain/blocks/PPoW-Blockchain-Block.js
#	src/common/mining-pools/miner/Miner-Pool-Settings.js
#	src/common/mining-pools/pool/pool-management/pool-work/Pool-Work.js
#	src/common/sockets/protocol/extend-socket/Node-Protocol.js
#	src/consts/const_global.js
#	src/node/menu/CLI-Menu.js
#	src/node/sockets/node-clients/service/discovery/fallbacks/fallback_nodes_list.js
#	src/node/sockets/node-server/express/Node-Express.js

+++ src/common/blockchain/interface-blockchain/blocks/Interface-Blockchain-Block-Data-Transactions.js
@@ -46,6 +46,20 @@ class InterfaceBlockchainBlockDataTransactions {
 
     }
 
+    markBlockDataTransactionsToBeInPending(){
+
+        if (this.pendingTransactionsWereIncluded)
+            return;
+
+        for (let i=0; i<this.transactions.length; i++) {
+            if (this.transactions[i].pendingTransactionsIncluded === undefined) this.transactions[i].pendingTransactionsIncluded = 0;
+            this.transactions[i].pendingTransactionsIncluded++;
+        }
+
+        this.pendingTransactionsWereIncluded = true;
+
+    }
+
     destroyBlockDataTransactions(){
 
         for (let i=0; i<this.transactions.length; i++) {
@@ -53,7 +67,9 @@ class InterfaceBlockchainBlockDataTransactions {
             if (this.pendingTransactionsWereIncluded)
                 this.transactions[i].pendingTransactionsIncluded--;
 
-            if ( Blockchain.blockchain.transactions.pendingQueue.findPendingTransaction(this.transactions[i].txId) === null )
+            if (this.pendingTransactionsWereIncluded <= 0) this.pendingTransactionsWereIncluded = undefined;
+
+            if ( !this.pendingTransactionsWereIncluded && Blockchain.blockchain.transactions.pendingQueue.findPendingTransaction(this.transactions[i].txId) === null )
                 this.transactions[i].destroyTransaction();
 
             this.transactions[i] = undefined;
@@ -181,6 +197,8 @@ class InterfaceBlockchainBlockDataTransactions {
 
                 this.transactions.push(transaction);
             }
+
+            this.transactionsLoaded = true;
         }
 
         return offset;
@@ -217,6 +235,16 @@ class InterfaceBlockchainBlockDataTransactions {
 
         return fee;
     }
+
+    freeTransactionsFromMemory(){
+
+        if (consts.SETTINGS.FREE_TRANSACTIONS_FROM_MEMORY_MAX_NUMBER <= 0) return;
+
+        this.destroyBlockDataTransactions(true);
+        delete this.transactionsLoaded;
+
+    }
+
 
 }
